+++ errors/types_generated.go
@@ -7,21 +7,22 @@ var (
 	BasicAuthCredentialsMissing  = Definitions[2]
 	Jwt                          = Definitions[3]
 	JwtTokenExpired              = Definitions[4]
-	JwtTokenInvalid              = Definitions[5]
-	JwtTokenMissing              = Definitions[6]
-	Oauth2                       = Definitions[7]
-	BetaRateLimiter              = Definitions[8]
-	BetaRateLimiterKey           = Definitions[9]
-	Saml2                        = Definitions[10]
-	Saml                         = Definitions[11]
-	InsufficientPermissions      = Definitions[12]
-	BackendOpenapiValidation     = Definitions[14]
-	BetaBackendRateLimitExceeded = Definitions[15]
-	BackendTimeout               = Definitions[16]
-	BetaBackendTokenRequest      = Definitions[17]
-	BackendUnhealthy             = Definitions[18]
-	Sequence                     = Definitions[20]
-	UnexpectedStatus             = Definitions[21]
+	JwtTokenInactive             = Definitions[5]
+	JwtTokenInvalid              = Definitions[6]
+	JwtTokenMissing              = Definitions[7]
+	Oauth2                       = Definitions[8]
+	BetaRateLimiter              = Definitions[9]
+	BetaRateLimiterKey           = Definitions[10]
+	Saml2                        = Definitions[11]
+	Saml                         = Definitions[12]
+	InsufficientPermissions      = Definitions[13]
+	BackendOpenapiValidation     = Definitions[15]
+	BetaBackendRateLimitExceeded = Definitions[16]
+	BackendTimeout               = Definitions[17]
+	BetaBackendTokenRequest      = Definitions[18]
+	BackendUnhealthy             = Definitions[19]
+	Sequence                     = Definitions[21]
+	UnexpectedStatus             = Definitions[22]
 )
 
 // typeDefinitions holds all related error definitions which are
@@ -36,6 +37,7 @@ var types = typeDefinitions{
 	"basic_auth_credentials_missing":   BasicAuthCredentialsMissing,
 	"jwt":                              Jwt,
 	"jwt_token_expired":                JwtTokenExpired,
+	"jwt_token_inactive":               JwtTokenInactive,
 	"jwt_token_invalid":                JwtTokenInvalid,
 	"jwt_token_missing":                JwtTokenMissing,
 	"oauth2":                           Oauth2,